Extract state write/read from DefaultInstantExecution
Because DefaultInstantExecution was getting large and had too many
responsibilities.

Leaving the orchestration and cache layout handling responsibility to
DefaultInstantExecution and moving the knowledge of how to write/read
the state to InstantExecutionState.
